+++ Classes/Utility/UrlUtility.php
@@ -10,24 +10,28 @@ use \T3v\T3vCore\Utility\AbstractHelper;
  */
 class UrlUtility extends AbstractUtility {
   /**
-   * Function to encode an URL.
+   * Encodes an URL.
    *
    * @param string $url The URL
    * @return string The encoded URL
    */
   public static function encodeUrl($url) {
+    $url = (string) $url;
+
     $url = urlencode($url);
 
     return $url;
   }
 
   /**
-   * Function to decode an URL.
+   * Decodes an URL.
    *
    * @param string $url The URL
    * @return string The decoded URL
    */
   public static function decodeUrl($url) {
+    $url = (string) $url;
+
     $url = urldecode($url);
 
     return $url;