+++ CHANGELOG.md
@@ -3,6 +3,12 @@ Changelog
 
 Notable changes will be documented in this file. The project adheres to [Semantic Versioning].
 
+Unreleased
+----------
+
+* Improved `MailService`
+* Improved Utility classes
+
 2.1.0
 -----